requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.92.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pydantic==1.10.5
PyJWT==2.6.0
pytest==7.2.1
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.25.0
StrEnum==0.4.9
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
